+++ client/components/orders/OrderListCard.jsx
@@ -8,15 +8,15 @@ import {
 } from 'material-ui/Card';
 
 import OrderListRow from './OrderListRow.jsx';
+import OrderDetail from './orderDetail';
 
 import styles from './orders.css';
 
-const OrderListCard = ({ order, orderDetail, onExpandChange }) => {
+const OrderListCard = ({ order, onExpandChange }) => {
 
-  const expanded = order.orderID === orderDetail.orderID;
   return (
     <Card
-      expanded={expanded}
+      expanded={order.expanded}
       onExpandChange={() => onExpandChange(order.orderID)}
       style={{
         borderBottom: '1px solid rgb(224, 224, 224)'
@@ -32,8 +32,11 @@ const OrderListCard = ({ order, orderDetail, onExpandChange }) => {
           height: '50px'
         }}
       />
-      <CardText expandable>
-        Foobar
+      <CardText
+        expandable
+        style={{ padding: 0 }}
+      >
+        <OrderDetail />
       </CardText>
     </Card >
   );
@@ -41,7 +44,6 @@ const OrderListCard = ({ order, orderDetail, onExpandChange }) => {
 
 OrderListCard.propTypes = {
   order: PropTypes.object.isRequired,
-  orderDetail: PropTypes.object.isRequired,
   onExpandChange: PropTypes.func.isRequired
 };